src/requirements.txt
Django==2.2
djangorestframework==3.10.3
django-cors-headers==3.1.1
psycopg2-binary==2.8.4
djangorestframework-simplejwt==4.3.0
wheel
uwsgi==2.0.18
django_cron==0.5.1
google-api-python-client==1.7.11
